requirements.txt
discord.py
python-dotenv
discord.py
python-dotenv
